requirements.txt
pandas>=2.0.0
numpy>=1.24.0
boto3>=1.26.0
python-dotenv>=1.0.0
pyspark>=3.5.0
mlflow>=2.11.0
scikit-learn>=1.3.0
pyarrow>=14.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
evidently>=0.4.0
prometheus-client>=0.19.0
great-expectations>=0.18.0
apache-airflow>=2.9.0
